requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyyaml
uritemplate
inflection
drf-spectacular
pillow